% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.m.R
\docType{data}
\name{data.m}
\alias{data.m}
\title{A data for 221 participants}
\format{
A data.frame contains 221 obs of 7 variables. The variables are:
\describe{
  \item{ID}{The ID of these 221 people.}
  \item{age}{A numeric.}
  \item{gender}{A two level factor: "1" or "2"}
  \item{CD8T}{A numeric.}
  \item{CD4T}{A numeric.}
  \item{x}{A numeric.}
  \item{y}{A numeric.}
}
}
\usage{
data.m
}
\description{
A data for 221 participants. The data were fabricated.
}
\keyword{datasets}
