% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E.result.R
\docType{data}
\name{E.result}
\alias{E.result}
\title{An \code{Emas} results data.}
\format{
A data.frame produced by \code{Emas} contains 2000 obs of 16 variables.
}
\usage{
E.result
}
\description{
An \code{Emas} results data for 2000 CpGs. The data were fabricated.
}
\keyword{datasets}
