% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mvalue.R
\docType{data}
\name{Mvalue}
\alias{Mvalue}
\title{A M-value matrix for 221 participants}
\format{
A matrix contains 221 obs of 2000 CpGs.
}
\usage{
Mvalue
}
\description{
A M-value matrix for 221 participants. The data were fabricated.
}
\keyword{datasets}
